+++ R/uniqueValues.R
@@ -2,7 +2,7 @@
 # Contributed by John Mount jmount@win-vector.com , ownership assigned to Win-Vector LLC.
 # Win-Vector LLC currently distributes this code without intellectual property indemnification, warranty, claim of fitness of purpose, or any other guarantee under a GPL3 license.
 
-#' @importFrom dplyr %>% ungroup select_ mutate group_by_ summarize
+#' @importFrom dplyr ungroup mutate summarize
 NULL
 
 
@@ -25,8 +25,8 @@ replyr_uniqueValues <- function(x,cname) {
   }
   replyr_private_value_n <- NULL # false binding for 'replyr_private_value_n' so name does not look unbound to CRAN check
   x %>% dplyr::ungroup() %>%
-    dplyr::select_(cname) %>% dplyr::mutate(replyr_private_value_n=1.0) %>%
-    dplyr::group_by_(cname) %>% dplyr::summarize(replyr_private_value_n=sum(replyr_private_value_n)) -> res
+    replyr_select(cname) %>% dplyr::mutate(replyr_private_value_n=1.0) %>%
+    replyr_select(cname) %>% dplyr::summarize(replyr_private_value_n=sum(replyr_private_value_n)) -> res
   # # Can't get rid of the warning on MySQL, even the following doesn't shut it up
   # suppressWarnings(
   #   # on mutate step in MySQL:  In .local(conn, statement, ...) : Decimal MySQL column 1 imported as numeric